chore: rename constants to plural, fix types
