+++ frontend/constants/intervals.ts
@@ -1,2 +1,1 @@
-export const SERVICE_CARD_STATUS_POLLING_INTERVAL = 5000;
-export const SERVICE_CARD_RPC_POLLING_INTERVAL = 5000;
+export const FIVE_SECONDS_INTERVAL = 5000;
