+++ frontend/constants/symbols.ts
